fix: Database.DPDT.Idris.return is now 'free'
There was some question over whether or not return should be

    return : a -> Private s a

or

    return : a -> Private 0 a

The answer is the second one.
And by making a minor change to the implementation of
Data.Rational.ratPlus, sequencing now works as expected.

The fixity of (>>=) was causing type unification to fail.
Consider:

    foo : Private s a -> Private s a
    foo p = do
      x <- p
      return x

This example could not compile before, but does now.
(Of course, this example was trivially solved with foo = id,
 but more complex examples couldn't be.)